libdnf-cli: Make tty.hpp a public header, move out of utils
The tty module seems widely needed, make it part of the libdnf::cli
namespace.
