+++ libdnf-cli/progressbar/download_progress_bar.cpp
@@ -19,13 +19,10 @@ along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 
 
 #include "libdnf-cli/progressbar/download_progress_bar.hpp"
-#include "libdnf-cli/utils/tty.hpp"
+#include "libdnf-cli/tty.hpp"
 
 #include <algorithm>
 #include <iomanip>
-
-
-using namespace libdnf::cli::utils;
 
 
 namespace libdnf::cli::progressbar {
