+++ microdnf/commands/advisory/advisory_info.cpp
@@ -22,11 +22,10 @@ along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 
 #include "../../context.hpp"
 
-#include "libdnf-cli/utils/tty.hpp"
-
 #include <libdnf-cli/output/advisoryinfo.hpp>
 #include <libdnf/rpm/package_query.hpp>
 
+#include <iostream>
 #include <filesystem>
 #include <fstream>
 #include <set>
